Move infrastructure code closer to tests for now
