+++ Cql/CoreTests/Tuples/StructuralValueTupleTests.cs
@@ -1,14 +1,17 @@
 ﻿#nullable enable
-using Hl7.Cql.Tuples;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
+using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text.Json;
+using Hl7.Cql.Abstractions.Infrastructure;
 
 namespace CoreTests.Tuples;
 
+#region Tests
+
 [TestClass]
-public class StructuralValueTupleBasicTests
+public partial class StructuralValueTupleBasicTests
 {
     [TestMethod]
     public void Constructor_ShouldSetProperties()
@@ -238,8 +241,54 @@ public class StructuralValueTupleBasicTests
         // Assert
         Assert.AreEqual(hashCodeA, hashCodeB);
     }
+}
 
+[TestClass]
+public partial class StructuralValueTupleSerializationTests
+{
+    [TestMethod]
+    public void TestJsonSerializationNested()
+    {
+        TupleAddress homeAddr =
+            new("Home", "Joe Street", "Springfield", "USA");
 
+        TupleAddress workAddr =
+            new("Work", "Sue Street", "Jumpville", "Canada");
+
+        TuplePerson person =
+            new("John", 10, [homeAddr, workAddr]);
+
+        var serializedJson = JsonSerializer.Serialize(person, new JsonSerializerOptions { WriteIndented = true });
+        Assert.AreEqual(
+            """
+            {
+              "Name": "John",
+              "Id": 10,
+              "Addresses": [
+                {
+                  "AddressType": "Home",
+                  "Street": "Joe Street",
+                  "City": "Springfield",
+                  "Country": "USA"
+                },
+                {
+                  "AddressType": "Work",
+                  "Street": "Sue Street",
+                  "City": "Jumpville",
+                  "Country": "Canada"
+                }
+              ]
+            }
+            """, serializedJson);
+    }
+}
+
+#endregion
+
+#region Generated C# as part of Library
+
+partial class StructuralValueTupleBasicTests
+{
     // Assume this is what the generated tuple will look like, just with a more generic name
     // [CompilerGenerated]
     public readonly struct TupleA(string? name = null, DateTime? dob = null)
@@ -265,11 +314,13 @@ public class StructuralValueTupleBasicTests
             1 => DOB,
             _ => throw new IndexOutOfRangeException()
         };
+
         int ITuple.Length => 2;
         string? IStructuralValue<string?>.Item1 => Name;
         DateTime? IStructuralValue<string?, DateTime?>.Item2 => DOB;
 
         StructuralValueMetadata IStructuralValue.GetMetadata() => Metadata;
+
         public static StructuralValueMetadata Metadata { get; } = new(
             [typeof(string), typeof(DateTime?)],
             ["Name", "DOB"]);
@@ -318,11 +369,13 @@ public class StructuralValueTupleBasicTests
             1 => DOB,
             _ => throw new IndexOutOfRangeException()
         };
+
         int ITuple.Length => 2;
         string? IStructuralValue<string?>.Item1 => Name;
         DateTime? IStructuralValue<string?, DateTime?>.Item2 => DOB;
 
         StructuralValueMetadata IStructuralValue.GetMetadata() => Metadata;
+
         public static StructuralValueMetadata Metadata { get; } = new(
             [typeof(string), typeof(DateTime?)],
             ["Name", "DOB"]);
@@ -344,48 +397,10 @@ public class StructuralValueTupleBasicTests
 
         public static bool operator !=(TupleB left, IStructuralValue<string?, DateTime?> right) => !(left == right);
     }
-
 }
 
-[TestClass]
-public class StructuralValueTupleSerializationTests
+partial class StructuralValueTupleSerializationTests
 {
-    [TestMethod]
-    public void TestJsonSerializationNested()
-    {
-        TupleAddress homeAddr =
-            new("Home", "Joe Street", "Springfield", "USA");
-
-        TupleAddress workAddr =
-            new("Work", "Sue Street", "Jumpville", "Canada");
-
-        TuplePerson person =
-            new("John", 10, [homeAddr, workAddr]);
-
-        var serializedJson = JsonSerializer.Serialize(person, new JsonSerializerOptions { WriteIndented = true });
-        Assert.AreEqual(
-            """
-            {
-              "Name": "John",
-              "Id": 10,
-              "Addresses": [
-                {
-                  "AddressType": "Home",
-                  "Street": "Joe Street",
-                  "City": "Springfield",
-                  "Country": "USA"
-                },
-                {
-                  "AddressType": "Work",
-                  "Street": "Sue Street",
-                  "City": "Jumpville",
-                  "Country": "Canada"
-                }
-              ]
-            }
-            """, serializedJson);
-    }
-
     // Assume this is what the generated tuple will look like, just with a more generic name
     // [CompilerGenerated]
     public readonly struct TuplePerson(string? name = null, int? id = null, TupleAddress[]? addresses = null)
@@ -397,13 +412,17 @@ public class StructuralValueTupleSerializationTests
         public int? Id { get; init; } = id;
         public TupleAddress[]? Addresses { get; init; } = addresses;
 
-        public TuplePerson(IStructuralValue<string?, int?, TupleAddress[]?> other) : this(other.Item1, other.Item2, other.Item3)
+        public TuplePerson(IStructuralValue<string?, int?, TupleAddress[]?> other) : this(
+            other.Item1, other.Item2, other.Item3)
         {
             if (other.GetMetadata().Signature != Metadata.Signature)
                 throw new ArgumentException("Incompatible metadata", nameof(other));
         }
 
-        public void Deconstruct(out string? name, out int? id, out TupleAddress[]? addresses) =>
+        public void Deconstruct(
+            out string? name,
+            out int? id,
+            out TupleAddress[]? addresses) =>
             (name, id, addresses) = (Name, Id, Addresses);
 
         object? ITuple.this[int index] => index switch
@@ -413,12 +432,14 @@ public class StructuralValueTupleSerializationTests
             2 => Addresses,
             _ => throw new IndexOutOfRangeException()
         };
+
         int ITuple.Length => 4;
         string? IStructuralValue<string?>.Item1 => Name;
         int? IStructuralValue<string?, int?>.Item2 => Id;
         TupleAddress[]? IStructuralValue<string?, int?, TupleAddress[]?>.Item3 => Addresses;
 
         StructuralValueMetadata IStructuralValue.GetMetadata() => Metadata;
+
         public static StructuralValueMetadata Metadata { get; } = new(
             [typeof(string), typeof(DateTime?)],
             ["Name", "DOB"]);
@@ -436,16 +457,18 @@ public class StructuralValueTupleSerializationTests
 
         public override int GetHashCode() => (AddressType: Name, Street: Id, City: Addresses).GetHashCode();
 
-        public static bool operator ==(TuplePerson left, IStructuralValue<string?, DateTime?> right) => left.Equals(right);
+        public static bool operator ==(TuplePerson left, IStructuralValue<string?, DateTime?> right) =>
+            left.Equals(right);
 
-        public static bool operator !=(TuplePerson left, IStructuralValue<string?, DateTime?> right) => !(left == right);
+        public static bool operator !=(TuplePerson left, IStructuralValue<string?, DateTime?> right) =>
+            !(left == right);
     }
-
 
 
     // Assume this is what the generated tuple will look like, just with a more generic name
     // [CompilerGenerated]
-    public readonly struct TupleAddress(string? addressType = null, string? street = null, string? city = null, string? country = null)
+    public readonly struct TupleAddress
+        (string? addressType = null, string? street = null, string? city = null, string? country = null)
         : IStructuralValue<string?, string?, string?, string?>,
           IStructuralValueMetadataProvider,
           IEquatable<IStructuralValue<string?, string?, string?, string?>>
@@ -455,13 +478,18 @@ public class StructuralValueTupleSerializationTests
         public string? City { get; init; } = city;
         public string? Country { get; init; } = country;
 
-        public TupleAddress(IStructuralValue<string?, string?, string?, string?> other) : this(other.Item1, other.Item2, other.Item3, other.Item4)
+        public TupleAddress(IStructuralValue<string?, string?, string?, string?> other) : this(
+            other.Item1, other.Item2, other.Item3, other.Item4)
         {
             if (other.GetMetadata().Signature != Metadata.Signature)
                 throw new ArgumentException("Incompatible metadata", nameof(other));
         }
 
-        public void Deconstruct(out string? addressType, out string? street, out string? city, out string? country) =>
+        public void Deconstruct(
+            out string? addressType,
+            out string? street,
+            out string? city,
+            out string? country) =>
             (addressType, street, city, country) = (AddressType, Street, City, Country);
 
         object? ITuple.this[int index] => index switch
@@ -472,13 +500,15 @@ public class StructuralValueTupleSerializationTests
             3 => Country,
             _ => throw new IndexOutOfRangeException()
         };
+
         int ITuple.Length => 4;
         string? IStructuralValue<string?>.Item1 => AddressType;
         string? IStructuralValue<string?, string?>.Item2 => Street;
-        string? IStructuralValue<string?, string?,string?>.Item3 => City;
-        string? IStructuralValue<string?, string?,string?,string?>.Item4 => Country;
+        string? IStructuralValue<string?, string?, string?>.Item3 => City;
+        string? IStructuralValue<string?, string?, string?, string?>.Item4 => Country;
 
         StructuralValueMetadata IStructuralValue.GetMetadata() => Metadata;
+
         public static StructuralValueMetadata Metadata { get; } = new(
             [typeof(string), typeof(DateTime?)],
             ["Name", "DOB"]);
@@ -496,8 +526,79 @@ public class StructuralValueTupleSerializationTests
 
         public override int GetHashCode() => (AddressType, Street, City, Country).GetHashCode();
 
-        public static bool operator ==(TupleAddress left, IStructuralValue<string?, DateTime?> right) => left.Equals(right);
+        public static bool operator ==(TupleAddress left, IStructuralValue<string?, DateTime?> right) =>
+            left.Equals(right);
 
-        public static bool operator !=(TupleAddress left, IStructuralValue<string?, DateTime?> right) => !(left == right);
+        public static bool operator !=(TupleAddress left, IStructuralValue<string?, DateTime?> right) =>
+            !(left == right);
     }
 }
+
+#endregion
+
+#region Runtime SDK
+
+public interface IStructuralValue : ITuple
+{
+    StructuralValueMetadata GetMetadata();
+}
+
+public interface IStructuralValue<out T1> : IStructuralValue
+{
+    T1 Item1 { get; }
+}
+
+public interface IStructuralValue<out T1, out T2> : IStructuralValue<T1>
+{
+    T2 Item2 { get; }
+}
+
+public interface IStructuralValue<out T1, out T2, out T3> : IStructuralValue<T1, T2>
+{
+    T3 Item3 { get; }
+}
+
+public interface IStructuralValue<out T1, out T2, out T3, out T4> : IStructuralValue<T1, T2, T3>
+{
+    T4 Item4 { get; }
+}
+
+public interface IStructuralValueMetadataProvider
+{
+    static abstract StructuralValueMetadata Metadata { get; }
+}
+
+public readonly record struct StructuralValueMetadata
+{
+    public StructuralValueMetadata(
+        Type[] ItemTypes,
+        string[] ItemNames)
+    {
+        this.ItemTypes = ItemTypes;
+        this.ItemNames = ItemNames;
+        Signature = string.Join("&", ItemTypes.Zip(ItemNames).Select(t => $"{t.Item1.ToCSharpString()} {t.Item2}"));
+    }
+
+    public Type[] ItemTypes { get; }
+
+    public string[] ItemNames { get; }
+
+    public string Signature { get; }
+
+    public void Deconstruct(
+        out Type[] itemTypes,
+        out string[] itemNames,
+        out string signature)
+    {
+        itemTypes = ItemTypes;
+        itemNames = ItemNames;
+        signature = Signature;
+    }
+    public override string ToString() => Signature;
+
+    public bool Equals(StructuralValueMetadata? other) =>
+        other is { } o &&
+        Signature == o.Signature;
+}
+
+#endregion
