+++ Cql/Cql.Abstractions/Cql.Abstractions.csproj
@@ -12,7 +12,6 @@
 
 	<ItemGroup>
 	  <ProjectReference Include="..\Iso8601\Iso8601.csproj" />
-		<InternalsVisibleTo Include="Hl7.Cql.Tuples" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="Hl7.Cql.Conversion" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="Hl7.Cql.Packager" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="Hl7.Cql.MultiPackager" Key="$(LibraryPKHash)" />
