--- Cql/Cql.Tuples/Cql.Tuples.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<Import Project="..\..\cql-sdk.props" />
-
-	<PropertyGroup>
-		<AssemblyName>Hl7.Cql.Tuples</AssemblyName>
-		<RootNamespace>Hl7.Cql.Tuples</RootNamespace>
-		<Product>CQL engine runtime</Product>
-		<Description>Support package for Hl7.Cql. Contains support infrastructure for using value tuples in CQL execution.</Description>
-	</PropertyGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\Cql.Abstractions\Cql.Abstractions.csproj" />
-	</ItemGroup>
-
-</Project>
